Relocated CW classes for PSR-4 compliance. Updated autoloader with changes.
